feat(api): add firecrawl v2 support with version-aware ui
